+++ part2/courseinfo/src/App.js
@@ -1,0 +1,29 @@
+import React from 'react'
+import Course from './components/Course'
+
+const App = () => {
+    const course = {
+        id: 1,
+        name: 'Half Stack application development',
+        parts: [
+            {
+                name: 'Fundamentals of React',
+                exercises: 10,
+                id: 1
+            },
+            {
+                name: 'Using props to pass data',
+                exercises: 7,
+                id: 2
+            },
+            {
+                name: 'State of a component',
+                exercises: 14,
+                id: 3
+            }
+        ]
+    }
+    return <Course course={course} />
+}
+
+export default App
