+++ part2/courseinfo/src/components/Header.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Header = ({ course }) => {
+    return (
+      <h1>{course.name}</h1>
+    )
+  }
+
+export default Header
