+++ part2/courseinfo/src/components/Part.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Part = (props) => {
+    return (
+        <p>
+            {props.part.name} {props.part.exercises}
+        </p>
+    )
+}
+
+export default Part
